% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condIndTest.R
\name{condIndTest}
\alias{condIndTest}
\title{Wrapper function for conditional independence tests.}
\usage{
condIndTest(Y, E, X, method = "KCI", alpha = 0.05, parsMethod = list(),
  verbose = FALSE)
}
\arguments{
\item{Y}{An n-dimensional vector or a matrix or dataframe with n rows and p columns.}

\item{E}{An n-dimensional vector or a matrix or dataframe with n rows and p columns.}

\item{X}{An n-dimensional vector or a matrix or dataframe with n rows and p columns.}

\item{verbose}{}
}
\value{

}
\description{
Tests the null hypothesis that Y and E are independent given X.
}
\examples{

# Example 1
set.seed(1)
n <- 100
Z <- rnorm(n)
X <- 4 + 2 * Z + rnorm(n)
Y <- 3 * X^2 + Z + rnorm(n)
test1 <- condIndTest(X,Y,Z, method = "KCI")
cat("These data come from a distribution, for which X and Y are NOT
cond. ind. given Z.")
cat(paste("The p-value of the test is: ", test1$pvalue))

# Example 2
set.seed(1)
Z <- rnorm(n)
X <- 4 + 2 * Z + rnorm(n)
Y <- 3 + Z + rnorm(n)
test2 <- condIndTest(X,Y,Z, method = "KCI")
cat("The data come from a distribution, for which X and Y are cond.
ind. given Z.")
cat(paste("The p-value of the test is: ", test2$pvalue))

}
