% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCI.R
\name{KCI}
\alias{KCI}
\title{Kernel conditional independence test.}
\usage{
KCI(Y, E, X, width = 0, alpha = 0.05, unbiased = FALSE, approx = TRUE,
  bootstrap = TRUE, nRepBs = 500, lambda = 0.001, thresh = 1e-05,
  numEig = length(Y), verbose = FALSE)
}
\arguments{
\item{Y}{A vector of length n or a matrix or dataframe with n rows and p columns.}

\item{E}{A vector of length n or a matrix or dataframe with n rows and p columns.}

\item{X}{A matrix or dataframe with n rows and p columns.}

\item{width}{Kernel width}

\item{alpha}{Significance level.  Defaults to 0.05.}

\item{verbose}{If \code{TRUE}, intermediate output is provided. Defaults to \code{FALSE}.}
}
\value{
A list with the following entries:
\itemize{
 \item \code{testStatistic} the statistic Tr(K_{(ddot{(X)}|Z)} * K_{(Y|Z)})
 \item \code{criticalValue} the critical point at the p-value equal to alpha;
  obtained by bootstrapping if \code{bootstrap = TRUE}, otherwise obtained by Gamma approximation..
 \item \code{pvalue} The p-value for the null hypothesis that Y and E are independent given X.
 It is obtained by bootstrapping if \code{bootstrap = TRUE}, otherwise obtained by Gamma approximation.
 }
}
\description{
Tests the null hypothesis that Y and E are independent given X.
}
\examples{
# Example 1
n <- 100
E <- rnorm(n)
X <- 4 + 2 * E + rnorm(n)
Y <- 3 * (X)^2 + rnorm(n)
KCI(Y, E, X)
KCI(Y, X, E)

}
