% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leveneAndWilcoxResidualDistributions.R
\name{leveneAndWilcoxResidualDistributions}
\alias{leveneAndWilcoxResidualDistributions}
\title{Levene and wilcoxon test to compare first and second moments of residual distributions}
\usage{
leveneAndWilcoxResidualDistributions(Y, predicted, E, verbose)
}
\arguments{
\item{Y}{An n-dimensional vector.}

\item{predicted}{An n-dimensional vector of predictions for Y.}

\item{E}{An n-dimensional vector. \code{E} needs to be a factor.}

\item{verbose}{Set to \code{TRUE} if output should be printed.}
}
\value{
A list with the p-value for the test.
}
\description{
Used as a subroutine in \code{InvariantResidualDistributionTest}
to test whether residual distribution remains invariant across different levels
of E.
}
